requirements.txt
python-fasthtml
lxml
requests
langchain
langchain-community
langchain-google-genai
beautifulsoup4
faiss-cpu